ring function fixed

+++ backend/db/migrations/20220620163931-create-booking.js
@@ -19,7 +19,7 @@ module.exports = {
         references: {model: 'Docks'}
       },
       totalCost: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC
       },
       startDate: {
         type: Sequelize.DATE
